+++ app/admin/payments.rb
@@ -12,7 +12,7 @@ ActiveAdmin.register Payment do
   #
   permit_params do
     permitted = [:user_id, :order_id, :source, :state, :amount_cents]
-    permitted << :organization_id if current_admin_user.is_root?
+    permitted << :organization_id if current_admin_user.is_root? || params[:action] == 'create'
     permitted
   end
 
@@ -21,6 +21,10 @@ ActiveAdmin.register Payment do
       myscope = super
       myscope = myscope.includes :organization, :user, :order
       myscope
+    end
+    def create
+      params[:payment][:organization_id] = current_admin_user.organization_id unless current_admin_user.is_root?
+      super
     end
   end
 
@@ -59,5 +63,72 @@ ActiveAdmin.register Payment do
       row(:created_at)
       row(:updated_at)
     end
+  end
+
+  form do |f|
+    columns do
+      column do
+        if current_admin_user.is_root?
+          f.inputs 'Admin' do
+            f.input :organization_id, as: :nested_select,
+                    minimum_input_length: 0,
+                    level_1: {
+                        attribute: :organization_id,
+                        fields: [:name],
+                        display_name: :name,
+                        minimum_input_length: 3,
+                        url: '\admin\organizations',
+                    },
+                    level_2: {
+                        attribute: :user_id,
+                        fields: [:email],
+                        display_name: :email,
+                        minimum_input_length: 3,
+                        url: '\admin\users',
+                    },
+                    level_3: {
+                        attribute: :order_id,
+                        fields: [:id],
+                        display_name: :start_on,
+                        url: '\admin\orders',
+                    }
+          end
+        else
+          f.inputs do
+            #tmp_params = current_admin_user.is_root? ? nil : { 'q[organization_id_equals]' => f.object.organization_id }
+            # f.input :user_id, as: :search_select, url: admin_users_path(tmp_params),
+            #         fields: [:email], display_name: 'email', minimum_input_length: 3,
+            #         order_by: 'email_asc', input_html: {class: 'after_user_selection'}
+            # f.input :order_id, as: :search_select, url: admin_orders_path( 'q[user_id_equals]' => f.object.user_id),
+            #         fields: [:start_on], display_name: 'id', minimum_input_length: 2,
+            #         order_by: 'id_asc'
+            # f.input :order, as: :select, collection: (f.object.user_id ? f.object.users.active : []), input_html: {class: 'orders_target'}
+            f.input :user_id, as: :nested_select,
+                    minimum_input_length: 0,
+                    level_1: {
+                        attribute: :user_id,
+                        fields: [:email],
+                        display_name: :email,
+                        minimum_input_length: 3,
+                        url: '\admin\users',
+                    },
+                    level_2: {
+                        attribute: :order_id,
+                        fields: [:id],
+                        display_name: :start_on,
+                        url: '\admin\orders',
+                    }
+          end
+        end
+      end
+      column do
+        f.inputs do
+          f.input :state
+          f.input :amount_cents
+        end
+      end
+    end
+
+    f.actions
   end
 end