+++ app/admin/subscriptions.rb
@@ -29,10 +29,11 @@ ActiveAdmin.register Subscription do
       if current_admin_user.is_root?
         f.input :organization
       end
-      f.input :category, collection: f.object.organization.categories
-      f.input :product, as: :select, collection: f.object.organization.products
+      f.input :category, collection: current_admin_user.categories
+      f.input :product, as: :select, collection: current_admin_user.products
       # Root admin could access all users so ability is not sufficient
-      f.input :user_id, as: :search_select, url: admin_users_path('q[organization_id_equals]' => f.object.organization_id),
+      tmp_params = current_admin_user.is_root? ? nil : { 'q[organization_id_equals]' => f.object.organization_id }
+      f.input :user_id, as: :search_select, url: admin_users_path(tmp_params),
               fields: [:email], display_name: 'email', minimum_input_length: 3,
               order_by: 'email_asc'
       f.input :code